// src/app/interfaces/player.ts
export class Player {
  _id: string;
  playerId: string;
  username: string;
  firstName: string;
  lastName: string;
  favoriteTeam: string;
  hatedTeam: string;
  atBatSong: string;
  dateOfBirth: Date;
  email: string;
  city: string;
  state: string;
  country: string;
  zipCode: string;
  modifiedDate: Date;
  createdDate: Date;
}